Added routes count in index + overhaul of all query and migrations

+++ database/migrations/2019_06_11_232200_create_api_table.php
@@ -14,7 +14,7 @@ class CreateApiTable extends Migration
     public function up()
     {
         Schema::create('api', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->unsignedInteger('id');
             $table->string('name');
         });
     }
