+++ database/migrations/2019_06_15_170447_create_routes_table.php
@@ -14,8 +14,8 @@ class CreateRoutesTable extends Migration
     public function up()
     {
         Schema::create('routes', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('api_id');
+            $table->unsignedInteger('id');
+            $table->unsignedInteger('api_id');
             $table->foreign('api_id')->references('id')->on('api')->onDelete('cascade');
             $table->string('method');
             $table->string('route');
